Add recording kind to fwd, cov, inv, stc filenames
If a dataset includes both, MEG and EEG data, we need a way to tell the
different forward models, noise covariance matrices, inverse operators,
and STCs apart.

In line with most of the other existing code, I'm using
config.get_kind() here (instead of config.ch_types), so currently, only
separation by KIND is possible and usability is limited. We will need to
address this at a later time – it's beyond the scope of this PR, and
currently not too urgent, I believe.