[MRG] BF: Don't alter MNE configuration (#142)
* BF: Call set_config only once

When processing multiple participants in parallel, one could encounter a
race condition if multiple processes were trying to write to the MNE
configuration file simultaneously. We now write the required setting
only once, at the beginning of the script, before forking any
sub-processes.

* Remove mne config setting

In fact, I believe we don't need this at all, since we're passing the
subjects_dir to all functions that require it anyway.