Rework ICA, test & add it to report
We're also writing a JSON file with useful ICA info; however I'm not
sure just HOW useful this file is, given that all information is
contains can also be retrieved from the ICA object / file directly.
In any case, in the future I would like to add an automated table
generator that provides an easy overview of the rejected ICs at a
single glance.