Fix incorrect step display in log output (#436)
* Fix incorrect step display in log output

The issue we're seeing (processing steps are not displayed correctly / are
outdated in the log) is due to the fact that we're using an environment
variable, runpy, parallelization, combined with Python's caching of
imports.

Afer fiddling with this for quite a while and reading LOTS of Stack Overflow
posts, I conclude that we cannot cleanly achieve our goal by setting an
environment variable in run.py. We are not seeing similar issues with all
the other variables we're using because they don't change over the course
of the pipeline run. However, the MNE_BIDS_STUDY_SCRIPT_PATH is
supposed to be updated with each new script that we call. This is why the
issue surfaces there.

My (interim?) solution is to write the path of the currently-executed script
to disk, read it back from there in config.py, and set the environment
variable accordingly – in config.py, too, and not in run.py as we currently
do on `main`.

This comes with the limitation that we can only have one pipeline instance
run at a time, but I guess this is okay for now. There are certainly ways to
resolve this and support multiple pipelines running in parallel, but I
think we need to discuss this.

WDYT?

* Update run.py

* Create MNE settings dir if it doesn't exist

* Fix incorrect step display in log output

The issue we're seeing (processing steps are not displayed correctly / are
outdated in the log) is due to the fact that we're using an environment
variable, runpy, parallelization, combined with Python's caching of
imports.

Afer fiddling with this for quite a while and reading LOTS of Stack Overflow
posts, I conclude that we cannot cleanly achieve our goal by setting an
environment variable in run.py. We are not seeing similar issues with all
the other variables we're using because they don't change over the course
of the pipeline run. However, the MNE_BIDS_STUDY_SCRIPT_PATH is
supposed to be updated with each new script that we call. This is why the
issue surfaces there.

My (interim?) solution is to write the path of the currently-executed script
to disk, read it back from there in config.py, and set the environment
variable accordingly – in config.py, too, and not in run.py as we currently
do on `main`.

This comes with the limitation that we can only have one pipeline instance
run at a time, but I guess this is okay for now. There are certainly ways to
resolve this and support multiple pipelines running in parallel, but I
think we need to discuss this.

WDYT?

* Revert "Fix incorrect step display in log output"

This reverts commit 783c7d2e948ed5c74c9802435360eb075dcf7f2e.

* Pass script path to failsafe_run decorator

* Remove cruft

* Fix doc build?