sanitize bug + compute maximal resamp + csp_quick
This was a very difficult commit. We need to reduce the running time for this step even if decimation is not enabled for the dataset. Because otherwise the circle CI would take something like 30 minutes to compute , which is not acceptable.

This is why I decimate by adding the csp_quick parameter. But this parameters introduces slights inacuracies for extrem time bins.

The problem is that by resampling or decimating, we change the limits of the epochs. for example  -5s, 5sbecomes -4.96,4.96, so time beens are no longer of the same size.

The solution would be to recompute the Tf object after decimation, but this would be way too much over engineered.

Let's keep it that way for the moment.