Use cleaned epochs for sensor-space decoding
We accidentally used the non-cleaned epochs for classification.
This issue probably snuck in during the recent set of refactorings.

@larsoner I've only checked the decoding scripts so far, because
this is what I'm currently working with. I'm not sure if other places
are affected as well. We should check the source-space scripts too.

Could you kindly take over if you find the time?